requirements.txt
streamlit
PyMuPDF
python-docx
google-genai
python-dotenv
plotly
pandas
fpdf2
